+++ src/Bromine/Trait/SpriteRenderTrait.cpp
@@ -4,8 +4,8 @@ namespace BromineEngine {
 
 DEFINE_TRAIT(SpriteRenderTrait, RenderTrait)
 
-SpriteRenderTrait::SpriteRenderTrait(const NodeID owner, ResourceID texture) : SpriteRenderTrait(owner, texture, Vec2d(1.0, 1.0)) {}
-SpriteRenderTrait::SpriteRenderTrait(const NodeID owner, ResourceID texture, Vec2d scale) : CONSTRUCT_TRAIT(SpriteRenderTrait, RenderTrait), texture(texture), _scale(scale) {}
+SpriteRenderTrait::SpriteRenderTrait(const NodeID owner, ResourceID texture) : SpriteRenderTrait(owner, texture, Vec2f(1.0, 1.0)) {}
+SpriteRenderTrait::SpriteRenderTrait(const NodeID owner, ResourceID texture, Vec2f scale) : CONSTRUCT_TRAIT(SpriteRenderTrait, RenderTrait), texture(texture), _scale(scale) {}
 
 void SpriteRenderTrait::render() {
 	// TODO: One of these is not like the other...
@@ -13,11 +13,11 @@ void SpriteRenderTrait::render() {
 	// Bromine::server.drawTexture(&getOwner().position(), &_scale, texture);
 }
 
-Vec2d& SpriteRenderTrait::scale() {
+Vec2f& SpriteRenderTrait::scale() {
 	return _scale;
 }
 
-Vec2d SpriteRenderTrait::scale() const {
+Vec2f SpriteRenderTrait::scale() const {
 	return _scale;
 }
